chore: move custom exception inside core module
