+++ application/rest-api-adapter/src/main/java/io/gebio/gebioback/rest/api/adapter/service/AuthenticationService.java
@@ -1,6 +1,6 @@
 package io.gebio.gebioback.rest.api.adapter.service;
 
-import io.gebio.gebioback.domain.exception.InvalidJwtToken;
+import io.gebio.gebioback.core.exception.InvalidJwtToken;
 import io.gebio.gebioback.domain.model.User;
 import io.gebio.gebioback.domain.port.in.UserFacade;
 import java.util.Optional;
